% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betatest.R
\name{betatest}
\alias{betatest}
\title{Hypothesis Test of the Effects of Idiosyncratic Factors}
\usage{
betatest(
  X,
  y,
  K,
  rmax = 8,
  method = "ic",
  weight = TRUE,
  type = "IC3",
  lambda.seq = NULL,
  alpha = 0.05
)
}
\arguments{
\item{X}{A list of the multi-modality data, each element is a data matrix of each group with dimension n*p_m.}

\item{y}{A response vector.}

\item{K}{The number of modalities.}

\item{rmax}{The maximum factor numbers of all modalities.}

\item{method}{The method used in the APM algorithm, default is ic, it can also be gap.}

\item{weight}{The weight of each projection matrix in the APM algorithm, default is TRUE. if weight = FALSE, then simply calculate the mean of all projection matrices.}

\item{type}{The method used in estimating the factor numbers in each modality initially, default is IC3.}

\item{lambda.seq}{A user supplied non-negative tuning parameters.}

\item{alpha}{The significance level used in hypothesis testing. The default is 0.05.}
}
\value{
A list containing the following components:
  \describe{
    \item{res}{A numeric indicator vector of length \eqn{1 + K + p}.
      The first element corresponds to the overall significance test,
      the next \eqn{K} elements correspond to the significance results
      for each modality, and the remaining \eqn{p} elements correspond
      to the significance results for each individual variable.}

    \item{pvals}{A numeric vector of p-values with the same length
      (\eqn{1 + K + p}) as \code{res}. Each element corresponds to the
      respective test in \code{res}.}
  }
}
\description{
Hypothesis Test of the Effects of Idiosyncratic Factors
}
\examples{
n=300
set.seed(2025)
K=4
p=1200
r_f=c(2,rep(2,K))
r_all=sum(r_f)
gamma1=0.5*rep(1,r_all)
pm=rep(p/K,K)
beta_0=c(rep(0,p))
# ------------------------#beta in the alternative with signal strength w rises from 0.05 to 0.2--------------
# beta_1=c(rep(0.05,2),rep(0,p/K-2),rep(0.05,2),rep(0,p/K-2),rep(0.05,2),rep(0,p/K-2),rep(0.05,2),rep(0,p/K-2))
# beta_2=c(rep(0.1,2),rep(0,p/K-2),rep(0.1,2),rep(0,p/K-2),rep(0.1,2),rep(0,p/K-2),rep(0.1,2),rep(0,p/K-2))
# beta_3=c(rep(0.15,2),rep(0,p/K-2),rep(0.15,2),rep(0,p/K-2),rep(0.15,2),rep(0,p/K-2),rep(0.15,2),rep(0,p/K-2))
# beta_4=c(rep(0.2,2),rep(0,p/K-2),rep(0.2,2),rep(0,p/K-2),rep(0.2,2),rep(0,p/K-2),rep(0.2,2),rep(0,p/K-2))
library(Matrix)
B_c<-matrix(rnorm(p*r_f[1]),nrow=p)
B=B_c
B_s<-list()
  for(k in 1:K){
    B_s[[k]]<-matrix(rnorm(pm[k]*r_f[k+1]),nrow=pm[k])
  }
block_diag <- as.matrix(bdiag(B_s[[1]], B_s[[2]]))
block_diag <- as.matrix(bdiag(block_diag, B_s[[3]]))
block_diag <- as.matrix(bdiag(block_diag, B_s[[4]]))
B=cbind(B_c,as.matrix(block_diag))
F<-matrix(rnorm(n*r_all),nrow=n)
V<-matrix(rnorm(n*p),nrow=n)
X=F\%*\%t(B)+V
ss=1
XX=list()
for(k in 1:K){
  XX[[k]]=X[,ss:(ss+pm[k]-1)]
  ss=ss+pm[k]
}
Y=F\%*\%gamma1+rnorm(n,0,1)
res <- betatest(XX,Y,K)
}
