% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{estimation}
\alias{estimation}
\title{Estimation function for the proposed model.}
\usage{
estimation(
  X,
  y,
  K,
  rmax = 8,
  method = "ic",
  weight = TRUE,
  type = "IC3",
  lambda.seq = NULL
)
}
\arguments{
\item{X}{A list of the multi-modality data, each element is a data matrix of each group with dimension n*p_m.}

\item{y}{A response vector.}

\item{K}{The number of modalities.}

\item{rmax}{The maximum factor numbers of all modalities.}

\item{method}{The method used in the APM algorithm, default is ic, it can also be gap.}

\item{weight}{The weight of each projection matrix in the APM algorithm, default is TRUE. if weight = FALSE, then simply calculate the mean of all projection matrices.}

\item{type}{The method used in estimating the factor numbers in each modality initially, default is IC3.}

\item{lambda.seq}{A user supplied non-negative tuning parameters.}
}
\value{
An object of class "estimation" containing the fitted model.
}
\description{
Estimation function for the proposed model.
}
\examples{
set.seed(1)
K=4
p=1200
library(Matrix)
r_f=c(2,rep(2,K))
r_all=sum(r_f)
pm=rep(p/K,K)
gamma1=rep(0.5,r_all)
beta=c(rep(0.5,2),rep(0,pm[1]-2),rep(0.5,2),rep(0,pm[2]-2),rep(0.5,2),rep(0,pm[3]-2),rep(0.5,2),rep(0,pm[4]-2))
B_c<-matrix(rnorm(p*r_f[1]),nrow=p)
B=B_c
B_s<-list()
for(k in 1:K){
  B_s[[k]]<-matrix(rnorm(pm[k]*r_f[k+1]),nrow=pm[k])
}

block_diag <- as.matrix(bdiag(B_s[[1]], B_s[[2]]))
block_diag <- as.matrix(bdiag(block_diag, B_s[[3]]))
block_diag <- as.matrix(bdiag(block_diag, B_s[[4]]))
B=cbind(B_c,as.matrix(block_diag))
n=ceiling(log(p)/(0.5^2/4^2))
F<-matrix(rnorm(n*r_all),nrow=n)
V<-matrix(rnorm(n*p),nrow=n)
X=F\%*\%t(B)+V
Y=F\%*\%gamma1+V\%*\%beta+rnorm(n,mean=0,sd=1)

ss=1
XX=list()
for(k in 1:K){
  XX[[k]]=X[,ss:(ss+pm[k]-1)]
  ss=ss+pm[k]
}
fit<-estimation(XX,Y,K,rmax=8,method='ic',weight=TRUE)
}
