% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gammatest.R
\name{gammatest}
\alias{gammatest}
\title{Hypothesis Test of the Effects of Common and Modality-specific Factors.}
\usage{
gammatest(
  X,
  y,
  K,
  ratio1,
  ratio2,
  rmax = 8,
  method = "ic",
  weight = TRUE,
  type = "IC3",
  lambda.seq = NULL,
  r_f = NULL,
  alpha = 0.05
)
}
\arguments{
\item{X}{A list of the multi-modality data, each element is a data matrix of each group with dimension n*p_m.}

\item{y}{A response vector.}

\item{K}{The number of modalities.}

\item{ratio1}{The exponent used in the sure screening step.}

\item{ratio2}{The exponent used in the ANOVA test step.}

\item{rmax}{The maximum number of factors across all modalities.}

\item{method}{The method used in the APM algorithm. The default is "ic"; it can also be "gap".}

\item{weight}{Logical. Whether to use weighted projection matrices in the APM algorithm.
The default is TRUE. If FALSE, the algorithm simply calculates the mean of all projection matrices.}

\item{type}{The method used to estimate the number of factors in each modality initially.
The default is "IC3".}

\item{lambda.seq}{A user-supplied sequence of non-negative tuning parameters.}

\item{r_f}{A vector specifying the number of factors in each modality. The default is NULL.}

\item{alpha}{The significance level used in hypothesis testing. The default is 0.05.}
}
\value{
A numeric vector containing the p-values testing the significance of
  the effects of common and modality-specific factors.
}
\description{
Hypothesis Test of the Effects of Common and Modality-specific Factors.
}
\examples{
p <- 1200
n=300
set.seed(1)
K=4
r_f=c(2,rep(2,K))
r_all=sum(r_f)
pm=rep(p/K,K)
beta=c(rep(1,2),rep(0,pm[1]-2),rep(1,2),rep(0,pm[2]-2),rep(1,2),rep(0,pm[3]-2),rep(1,2),rep(0,pm[4]-2))
library(Matrix)
B_c<-matrix(rnorm(p*r_f[1],sd=1),nrow=p)
B=B_c
B_s<-list()
for(k in 1:K){
  B_s[[k]]<-matrix(rnorm(pm[k]*r_f[k+1],sd=1),nrow=pm[k])
}
block_diag <- as.matrix(bdiag(B_s[[1]], B_s[[2]]))
block_diag <- as.matrix(bdiag(block_diag, B_s[[3]]))
block_diag <- as.matrix(bdiag(block_diag, B_s[[4]]))
B=cbind(B_c,as.matrix(block_diag))
gamma1=c(rep(2,r_all))
set.seed(1)
F<-matrix(rnorm(n*r_all),nrow=n)
V<-matrix(rnorm(n*p),nrow=n)
X=F\%*\%t(B)+V #'design matrix
y=V\%*\%beta+rnorm(n,mean=0,sd=1)
y1=F\%*\%gamma1+y

ss=1
XX=list()
for(k in 1:K){
  XX[[k]]=X[,ss:(ss+pm[k]-1)]
  ss=ss+pm[k]
}
res=gammatest(XX,y,K,ratio1=0.9,ratio2=0.6,rmax=8,method='ic',weight=TRUE,lambda.seq=NULL,r_f=NULL,alpha=0.05,type="IC3")
}
