% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.estimation.R
\name{predict.estimation}
\alias{predict.estimation}
\title{Prediction method for estimation objects}
\usage{
\method{predict}{estimation}(object, newdata, ...)
}
\arguments{
\item{object}{An object of class "estimation" (returned by estimation()).}

\item{newdata}{A data frame containing the new predictor values.}
}
\value{
A numeric vector of predicted values.
}
\description{
Defines the predict() S3 method for objects of class "estimation".
}
\examples{

set.seed(1)
K=4
p=1200
library(Matrix)
r_f=c(2,rep(2,K))
r_all=sum(r_f)
pm=rep(p/K,K)
gamma1=rep(0.5,r_all)#'c(0,0.3,0.5,0.5,0.5,0,0,0,0,0.5)
beta=c(rep(0.5,2),rep(0,pm[1]-2),rep(0.5,2),rep(0,pm[2]-2),rep(0.5,2),rep(0,pm[3]-2),rep(0.5,2),rep(0,pm[4]-2))
B_c<-matrix(rnorm(p*r_f[1]),nrow=p)
B=B_c
B_s<-list()
for(k in 1:K){
  B_s[[k]]<-matrix(rnorm(pm[k]*r_f[k+1]),nrow=pm[k])
}

block_diag <- as.matrix(bdiag(B_s[[1]], B_s[[2]]))
block_diag <- as.matrix(bdiag(block_diag, B_s[[3]]))
block_diag <- as.matrix(bdiag(block_diag, B_s[[4]]))
B=cbind(B_c,as.matrix(block_diag))
n=ceiling(log(p)/(0.5^2/4^2))
F<-matrix(rnorm(n*r_all),nrow=n)
V<-matrix(rnorm(n*p),nrow=n)
X=F\%*\%t(B)+V
Y=F\%*\%gamma1+V\%*\%beta+rnorm(n,mean=0,sd=1)

ss=1
XX=list()
for(k in 1:K){
  XX[[k]]=X[,ss:(ss+pm[k]-1)]
  ss=ss+pm[k]
}
fit<-estimation(XX,Y,K,rmax=8,method='ic',weight=TRUE)
n_test <- 100
F_test <- matrix(rnorm(n_test * r_all), nrow = n_test)
V_test <- matrix(rnorm(n_test * p), nrow = n_test)
X_test <- F_test \%*\% t(B) + V_test
Y_test <- F_test\%*\%gamma1+V_test\%*\%beta+rnorm(n_test,mean=0,sd=1)
y_pred <- predict(fit,X_test)
}
